+++ etc/gunicorn.conf.py
@@ -1,0 +1,7 @@
+# The settings can be overwritten by mounting a differnt file into the docker container
+access_logfile = "-"
+bind = ":8000"
+capture_output = True
+log_file = "-"
+timeout = 60
+workers = 2