+++ Changelog.md
@@ -5,6 +5,7 @@ Please see the [Readme](Readme.md#Update]) on what commands to run after an upda
 ## v0.2.12 - Unreleased
 
  * Add `SENTRY_TRACES_SAMPLE_RATE` for the sentry integration (https://docs.sentry.io/platforms/python/guides/django/performance/).
+ * The gunicorn config is now in `etc/gunicorn.conf.py`, which can be overwritten by mounting a different file into the container
 
 ## v0.2.11